+++ src/StuffyHelper.Contracts/Models/PurchaseUsageShortEntry.cs
@@ -26,6 +26,6 @@ namespace StuffyHelper.Contracts.Models
         /// <summary>
         /// Amount of the purchase that was used
         /// </summary>
-        public double Amount { get; init; }
+        public long Amount { get; init; }
     }
 }
